+++ lib/dartapi_core.dart
@@ -8,4 +8,8 @@ export 'src/core/logger.dart';
 export 'src/core/serializable.dart';
 export 'src/core/global_exception_handler.dart';
 
+export 'src/openapi/security_scheme.dart';
+export 'src/openapi/openapi_generator.dart';
+export 'src/openapi/docs_controller.dart';
+
 export 'src/utils/utils.dart';